requirements.txt
openpyxl==2.5.9
httplib2==0.11.3
numpy==1.16.2
sendgrid==5.3.0
pandas==0.24.2
PyQt5==5.13.2
google_api_python_client==1.7.11
beautifulsoup4==4.8.1
python_dateutil==2.8.1
